Struct google_notebooks1::api::ExecutionTemplate

pub struct ExecutionTemplate {
    pub accelerator_config: Option<SchedulerAcceleratorConfig>,
    pub container_image_uri: Option<String>,
    pub input_notebook_file: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub master_type: Option<String>,
    pub output_notebook_folder: Option<String>,
    pub parameters: Option<String>,
    pub params_yaml_file: Option<String>,
    pub scale_tier: Option<String>,
}

The description a notebook execution workload.

This type is not used in any activity, and only used as part of another schema.

Fields
accelerator_config: Option<SchedulerAcceleratorConfig>

Configuration (count and accelerator type) for hardware running notebook execution.

container_image_uri: Option<String>

Container Image URI to a DLVM Example: 'gcr.io/deeplearning-platform-release/base-cu100' More examples can be found at: https://cloud.google.com/ai-platform/deep-learning-containers/docs/choosing-container

input_notebook_file: Option<String>

Path to the notebook file to execute. Must be in a Google Cloud Storage bucket. Format: gs://{project_id}/{folder}/{notebook_file_name} Ex: gs://notebook_user/scheduled_notebooks/sentiment_notebook.ipynb

labels: Option<HashMap<String, String>>

Labels for execution. If execution is scheduled, a field included will be 'nbs-scheduled'. Otherwise, it is an immediate execution, and an included field will be 'nbs-immediate'. Use fields to efficiently index between various types of executions.

master_type: Option<String>

Specifies the type of virtual machine to use for your training job's master worker. You must specify this field when scaleTier is set to CUSTOM. You can use certain Compute Engine machine types directly in this field. The following types are supported: - n1-standard-4 - n1-standard-8 - n1-standard-16 - n1-standard-32 - n1-standard-64 - n1-standard-96 - n1-highmem-2 - n1-highmem-4 - n1-highmem-8 - n1-highmem-16 - n1-highmem-32 - n1-highmem-64 - n1-highmem-96 - n1-highcpu-16 - n1-highcpu-32 - n1-highcpu-64 - n1-highcpu-96 Alternatively, you can use the following legacy machine types: - standard - large_model - complex_model_s - complex_model_m - complex_model_l - standard_gpu - complex_model_m_gpu - complex_model_l_gpu - standard_p100 - complex_model_m_p100 - standard_v100 - large_model_v100 - complex_model_m_v100 - complex_model_l_v100 Finally, if you want to use a TPU for training, specify cloud_tpu in this field. Learn more about the [special configuration options for training with TPU.

output_notebook_folder: Option<String>

Path to the notebook folder to write to. Must be in a Google Cloud Storage bucket path. Format: gs://{project_id}/{folder} Ex: gs://notebook_user/scheduled_notebooks

parameters: Option<String>

Parameters used within the 'input_notebook_file' notebook.

params_yaml_file: Option<String>

Parameters to be overridden in the notebook during execution. Ref https://papermill.readthedocs.io/en/latest/usage-parameterize.html on how to specifying parameters in the input notebook and pass them here in an YAML file. Ex: gs://notebook_user/scheduled_notebooks/sentiment_notebook_params.yaml

scale_tier: Option<String>

Required. Scale tier of the hardware used for notebook execution.