Struct google_notebooks1::api::Instance

pub struct Instance {
    pub accelerator_config: Option<AcceleratorConfig>,
    pub boot_disk_size_gb: Option<String>,
    pub boot_disk_type: Option<String>,
    pub container_image: Option<ContainerImage>,
    pub create_time: Option<String>,
    pub custom_gpu_driver_path: Option<String>,
    pub data_disk_size_gb: Option<String>,
    pub data_disk_type: Option<String>,
    pub disk_encryption: Option<String>,
    pub disks: Option<Vec<Disk>>,
    pub install_gpu_driver: Option<bool>,
    pub instance_owners: Option<Vec<String>>,
    pub kms_key: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub machine_type: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub no_proxy_access: Option<bool>,
    pub no_public_ip: Option<bool>,
    pub no_remove_data_disk: Option<bool>,
    pub post_startup_script: Option<String>,
    pub proxy_uri: Option<String>,
    pub service_account: Option<String>,
    pub service_account_scopes: Option<Vec<String>>,
    pub shielded_instance_config: Option<ShieldedInstanceConfig>,
    pub state: Option<String>,
    pub subnet: Option<String>,
    pub tags: Option<Vec<String>>,
    pub update_time: Option<String>,
    pub upgrade_history: Option<Vec<UpgradeHistoryEntry>>,
    pub vm_image: Option<VmImage>,
}

The definition of a notebook instance.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations instances create projects (request)

locations instances get projects (response)

Fields
accelerator_config: Option<AcceleratorConfig>

The hardware accelerator used on this instance. If you use accelerators, make sure that your configuration has enough vCPUs and memory to support the machine_type you have selected.

boot_disk_size_gb: Option<String>

Input only. The size of the boot disk in GB attached to this instance, up to a maximum of 64000 GB (64 TB). The minimum recommended value is 100 GB. If not specified, this defaults to 100.

boot_disk_type: Option<String>

Input only. The type of the boot disk attached to this instance, defaults to standard persistent disk (PD_STANDARD).

container_image: Option<ContainerImage>

Use a container image to start the notebook instance.

create_time: Option<String>

Output only. Instance creation time.

custom_gpu_driver_path: Option<String>

Specify a custom Cloud Storage path where the GPU driver is stored. If not specified, we'll automatically choose from official GPU drivers.

data_disk_size_gb: Option<String>

Input only. The size of the data disk in GB attached to this instance, up to a maximum of 64000 GB (64 TB). You can choose the size of the data disk based on how big your notebooks and data are. If not specified, this defaults to 100.

data_disk_type: Option<String>

Input only. The type of the data disk attached to this instance, defaults to standard persistent disk (PD_STANDARD).

disk_encryption: Option<String>

Input only. Disk encryption method used on the boot and data disks, defaults to GMEK.

disks: Option<Vec<Disk>>

Output only. Attached disks to notebook instance.

install_gpu_driver: Option<bool>

Whether the end user authorizes Google Cloud to install GPU driver on this instance. If this field is empty or set to false, the GPU driver won't be installed. Only applicable to instances with GPUs.

instance_owners: Option<Vec<String>>

Input only. The owner of this instance after creation. Format: alias@example.com Currently supports one owner only. If not specified, all of the service account users of your VM instance's service account can use the instance.

kms_key: Option<String>

Input only. The KMS key used to encrypt the disks, only applicable if disk_encryption is CMEK. Format: projects/{project_id}/locations/{location}/keyRings/{key_ring_id}/cryptoKeys/{key_id} Learn more about using your own encryption keys.

labels: Option<HashMap<String, String>>

Labels to apply to this instance. These can be later modified by the setLabels method.

machine_type: Option<String>

Required. The Compute Engine machine type of this instance.

metadata: Option<HashMap<String, String>>

Custom metadata to apply to this instance.

name: Option<String>

Output only. The name of this notebook instance. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

network: Option<String>

The name of the VPC that this instance is in. Format: projects/{project_id}/global/networks/{network_id}

no_proxy_access: Option<bool>

If true, the notebook instance will not register with the proxy.

no_public_ip: Option<bool>

If true, no public IP will be assigned to this instance.

no_remove_data_disk: Option<bool>

Input only. If true, the data disk will not be auto deleted when deleting the instance.

post_startup_script: Option<String>

Path to a Bash script that automatically runs after a notebook instance fully boots up. The path must be a URL or Cloud Storage path (gs://path-to-file/file-name).

proxy_uri: Option<String>

Output only. The proxy endpoint that is used to access the Jupyter notebook.

service_account: Option<String>

The service account on this instance, giving access to other Google Cloud services. You can use any service account within the same project, but you must have the service account user permission to use the instance. If not specified, the Compute Engine default service account is used.

service_account_scopes: Option<Vec<String>>

Optional. The URIs of service account scopes to be included in Compute Engine instances. If not specified, the following scopes are defined: - https://www.googleapis.com/auth/cloud-platform - https://www.googleapis.com/auth/userinfo.email If not using default scopes, you need at least: https://www.googleapis.com/auth/compute

shielded_instance_config: Option<ShieldedInstanceConfig>

Optional. Shielded VM configuration. [Images using supported Shielded VM features] (https://cloud.google.com/compute/docs/instances/modifying-shielded-vm).

state: Option<String>

Output only. The state of this instance.

subnet: Option<String>

The name of the subnet that this instance is in. Format: projects/{project_id}/regions/{region}/subnetworks/{subnetwork_id}

tags: Option<Vec<String>>

Optional. The Compute Engine tags to add to runtime (see Tagging instances).

update_time: Option<String>

Output only. Instance update time.

upgrade_history: Option<Vec<UpgradeHistoryEntry>>

The upgrade history of this instance.

vm_image: Option<VmImage>

Use a Compute Engine VM image to start the notebook instance.