Struct google_notebooks1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the AIPlatformNotebooks hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_notebooks1 as notebooks1;
 
use std::default::Default;
use oauth2;
use notebooks1::AIPlatformNotebooks;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = AIPlatformNotebooks::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_environments_create(...)`, `locations_environments_delete(...)`, `locations_environments_get(...)`, `locations_environments_list(...)`, `locations_executions_create(...)`, `locations_executions_delete(...)`, `locations_executions_get(...)`, `locations_executions_list(...)`, `locations_get(...)`, `locations_instances_create(...)`, `locations_instances_delete(...)`, `locations_instances_get(...)`, `locations_instances_get_iam_policy(...)`, `locations_instances_get_instance_health(...)`, `locations_instances_is_upgradeable(...)`, `locations_instances_list(...)`, `locations_instances_register(...)`, `locations_instances_report(...)`, `locations_instances_reset(...)`, `locations_instances_set_accelerator(...)`, `locations_instances_set_iam_policy(...)`, `locations_instances_set_labels(...)`, `locations_instances_set_machine_type(...)`, `locations_instances_start(...)`, `locations_instances_stop(...)`, `locations_instances_test_iam_permissions(...)`, `locations_instances_upgrade(...)`, `locations_instances_upgrade_internal(...)`, `locations_list(...)`, `locations_operations_cancel(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_schedules_create(...)`, `locations_schedules_delete(...)`, `locations_schedules_get(...)`, `locations_schedules_list(...)` and `locations_schedules_trigger(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_environments_create(
&self,
request: Environment,
parent: &str
) -> ProjectLocationEnvironmentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Environment.

Arguments

request - No description provided.

parent - Required. Format: projects/{project_id}/locations/{location}

pub fn locations_environments_delete(
&self,
name: &str
) -> ProjectLocationEnvironmentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Environment.

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/environments/{environment_id}

pub fn locations_environments_get(
&self,
name: &str
) -> ProjectLocationEnvironmentGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Environment.

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/environments/{environment_id}

pub fn locations_environments_list(
&self,
parent: &str
) -> ProjectLocationEnvironmentListCall<'a, C>

Create a builder to help you perform the following task:

Lists environments in a project.

Arguments

parent - Required. Format: projects/{project_id}/locations/{location}

pub fn locations_executions_create(
&self,
request: Execution,
parent: &str
) -> ProjectLocationExecutionCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Scheduled Notebook in a given project and location.

Arguments

request - No description provided.

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_executions_delete(
&self,
name: &str
) -> ProjectLocationExecutionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes execution

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/executions/{execution_id}

pub fn locations_executions_get(
&self,
name: &str
) -> ProjectLocationExecutionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of executions

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/schedules/{execution_id}

pub fn locations_executions_list(
&self,
parent: &str
) -> ProjectLocationExecutionListCall<'a, C>

Create a builder to help you perform the following task:

Lists executions in a given project and location

Arguments

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_instances_create(
&self,
request: Instance,
parent: &str
) -> ProjectLocationInstanceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Instance in a given project and location.

Arguments

request - No description provided.

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_instances_delete(
&self,
name: &str
) -> ProjectLocationInstanceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Instance.

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_get(
&self,
name: &str
) -> ProjectLocationInstanceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single Instance.

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationInstanceGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_instances_get_instance_health(
&self,
name: &str
) -> ProjectLocationInstanceGetInstanceHealthCall<'a, C>

Create a builder to help you perform the following task:

Check if a notebook instance is healthy.

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_is_upgradeable(
&self,
notebook_instance: &str
) -> ProjectLocationInstanceIsUpgradeableCall<'a, C>

Create a builder to help you perform the following task:

Check if a notebook instance is upgradable.

Arguments

notebookInstance - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_list(
&self,
parent: &str
) -> ProjectLocationInstanceListCall<'a, C>

Create a builder to help you perform the following task:

Lists instances in a given project and location.

Arguments

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_instances_register(
&self,
request: RegisterInstanceRequest,
parent: &str
) -> ProjectLocationInstanceRegisterCall<'a, C>

Create a builder to help you perform the following task:

Registers an existing legacy notebook instance to the Notebooks API server. Legacy instances are instances created with the legacy Compute Engine calls. They are not manageable by the Notebooks API out of the box. This call makes these instances manageable by the Notebooks API.

Arguments

request - No description provided.

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_instances_report(
&self,
request: ReportInstanceInfoRequest,
name: &str
) -> ProjectLocationInstanceReportCall<'a, C>

Create a builder to help you perform the following task:

Allows notebook instances to report their latest instance information to the Notebooks API server. The server will merge the reported information to the instance metadata store. Do not use this method directly.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_reset(
&self,
request: ResetInstanceRequest,
name: &str
) -> ProjectLocationInstanceResetCall<'a, C>

Create a builder to help you perform the following task:

Resets a notebook instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_set_accelerator(
&self,
request: SetInstanceAcceleratorRequest,
name: &str
) -> ProjectLocationInstanceSetAcceleratorCall<'a, C>

Create a builder to help you perform the following task:

Updates the guest accelerators of a single Instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationInstanceSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_instances_set_labels(
&self,
request: SetInstanceLabelsRequest,
name: &str
) -> ProjectLocationInstanceSetLabelCall<'a, C>

Create a builder to help you perform the following task:

Replaces all the labels of an Instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_set_machine_type(
&self,
request: SetInstanceMachineTypeRequest,
name: &str
) -> ProjectLocationInstanceSetMachineTypeCall<'a, C>

Create a builder to help you perform the following task:

Updates the machine type of a single Instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_start(
&self,
request: StartInstanceRequest,
name: &str
) -> ProjectLocationInstanceStartCall<'a, C>

Create a builder to help you perform the following task:

Starts a notebook instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_stop(
&self,
request: StopInstanceRequest,
name: &str
) -> ProjectLocationInstanceStopCall<'a, C>

Create a builder to help you perform the following task:

Stops a notebook instance.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationInstanceTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_instances_upgrade(
&self,
request: UpgradeInstanceRequest,
name: &str
) -> ProjectLocationInstanceUpgradeCall<'a, C>

Create a builder to help you perform the following task:

Upgrades a notebook instance to the latest version.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_instances_upgrade_internal(
&self,
request: UpgradeInstanceInternalRequest,
name: &str
) -> ProjectLocationInstanceUpgradeInternalCall<'a, C>

Create a builder to help you perform the following task:

Allows notebook instances to call this endpoint to upgrade themselves. Do not use this method directly.

Arguments

request - No description provided.

name - Required. Format: projects/{project_id}/locations/{location}/instances/{instance_id}

pub fn locations_operations_cancel(
&self,
request: CancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_schedules_create(
&self,
request: Schedule,
parent: &str
) -> ProjectLocationScheduleCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Scheduled Notebook in a given project and location.

Arguments

request - No description provided.

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_schedules_delete(
&self,
name: &str
) -> ProjectLocationScheduleDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes schedule and all underlying jobs

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/schedules/{schedule_id}

pub fn locations_schedules_get(
&self,
name: &str
) -> ProjectLocationScheduleGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of schedule

Arguments

name - Required. Format: projects/{project_id}/locations/{location}/schedules/{schedule_id}

pub fn locations_schedules_list(
&self,
parent: &str
) -> ProjectLocationScheduleListCall<'a, C>

Create a builder to help you perform the following task:

Lists schedules in a given project and location.

Arguments

parent - Required. Format: parent=projects/{project_id}/locations/{location}

pub fn locations_schedules_trigger(
&self,
request: TriggerScheduleRequest,
name: &str
) -> ProjectLocationScheduleTriggerCall<'a, C>

Create a builder to help you perform the following task:

Triggers execution of an existing schedule.

Arguments

request - No description provided.

name - Required. Format: parent=projects/{project_id}/locations/{location}/schedules/{schedule_id}

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.